// src/athlete/body-composition/dto/get-history-body-composition.dto.ts
import { GenderEnum } from 'src/user/enums/gender.enum';

export class getHistoryBodyCompositionDto {
  name: string;
  weight_class: string;
  sex: GenderEnum;
  test_date_time: string;
  height: number;
  weight: number;
  lower_limit_weight_normal_range: number;
  upper_limit_weight_normal_range: number;
  bmi_body_mass_index: number;
  lower_limit_bmi_normal_range: number;
  upper_limit_bmi_normal_range: number;
  bmr_basal_metabolic_rate: number;
  lower_limit_bmr_normal_range: number;
  upper_limit_bmr_normal_range: number;
  slm_soft_lean_mass: number;
  lower_limit_slm_normal_range: number;
  upper_limit_slm_normal_range: number;
  tbw_total_body_water: number;
  lower_limit_tbw_normal_range: number;
  upper_limit_tbw_normal_range: number;
  protein: number;
  lower_limit_protein_normal_range: number;
  upper_limit_protein_normal_range: number;
  minerals: number;
  visceral_fat: number;
  lower_limit_visceral_fat_of_abdomen_normal_range: number;
  upper_limit_visceral_fat_of_abdomen_normal_range: number;
  subcutaneous_fat: number;
  lower_limit_subcutaneous_fat_of_abdomen_normal_range: number;
  upper_limit_subcutaneous_fat_of_abdomen_normal_range: number;
  icw_intracellular_water: number;
  lower_limit_icw_normal_range: number;
  upper_limit_icw_normal_range: number;
  ecw_extracellular_water: number;
  lower_limit_ecw_normal_range: number;
  upper_limit_ecw_normal_range: number;
  smm_skeletal_muscle_mass: number;
  lower_limit_smm_normal_range: number;
  upper_limit_smm_normal_range: number;
  pbf_percent_body_fat: number;
  lower_limit_pbf_normal_range: number;
  upper_limit_pbf_normal_range: number;
  ffmi_fat_free_mass_index: number;
  fmi_fat_mass_index: number;
  smi: number;
  bfm_body_fat_mass: number;
  lower_limit_bfm_normal_range: number;
  upper_limit_bfm_normal_range: number;
}